# proteccio-backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Proteccio API
  version: 2.2.3
  description: |
    Proteccio project management platform API documentation.

    ## Authentication
    This API uses **session-based authentication** with **CSRF token protection**:

    1. **Session Cookie**: Automatically set after login
    2. **CSRF Token**: Required for all state-changing operations (POST, PUT, DELETE)
       - Obtain token from `GET /csrf-token` endpoint
       - Include in `X-CSRF-Token` header for subsequent requests

    ## Authorization Flow
    ```
    1. POST /auth/login → Receive session cookie
    2. GET /csrf-token → Receive CSRF token
    3. Use session + CSRF token for protected endpoints
    ```

    ## Response Format
    All endpoints return a standardized `ServerResponse<T>` object:
    ```json
    {
      "done": true,
      "body": { ... },
      "title": "Optional title",
      "message": "Optional message"
    }
    ```

  contact:
    name: Proteccio Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Development server
...
security:
  - sessionAuth: []
    csrfToken: []
paths:
...
  /tasks/{id}:
    put:
      operationId: putTasksById
      tags:
        - Tasks
      summary: Update task
      description: Update an existing task's properties
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ServerResponse'
                  - type: object
                    properties:
                      body:
                        $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: Task not found

    delete:
      operationId: deleteTasksById
      tags:
        - Tasks
      summary: Delete task
      description: Permanently delete a task
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
              example:
                done: true
                body: null
                message: "Task deleted successfully"
        '404':
          description: Task not found
  /tasks/project/{id}:
    get:
      operationId: getTasksProjectById
      tags:
        - Tasks
      summary: Get tasks by project
      description: Retrieve all tasks for a specific project
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ServerResponse'
                  - type: object
                    properties:
                      body:
                        type: array
                        items:
                          $ref: '#/components/schemas/Task'
...
  /projects/{id}:
    get:
      operationId: getProjectsById
      tags:
        - Projects
      summary: Get project by ID
      description: Retrieve detailed information about a specific project
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Project details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ServerResponse'
                  - type: object
                    properties:
                      body:
                        $ref: '#/components/schemas/Project'
        '404':
          description: Project not found

    put:
      operationId: putProjectsById
      tags:
        - Projects
      summary: Update project
      description: Update an existing project's properties (requires project manager role)
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ServerResponse'
                  - type: object
                    properties:
                      body:
                        $ref: '#/components/schemas/Project'
        '403':
          description: Insufficient permissions (requires project manager role)

    delete:
      operationId: deleteProjectsById
      tags:
        - Projects
      summary: Delete project
      description: Permanently delete a project (requires team owner or admin role)
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
              example:
                done: true
                body: null
                message: "Project deleted successfully"
        '403':
          description: Insufficient permissions
...
  /statuses/{id}:
    get:
      operationId: getStatusesById
      tags:
        - Statuses
      summary: Get status by ID
      description: Returns a specific status record for a project context.
      parameters:
        - $ref: '#/components/parameters/idParam'
        - $ref: '#/components/parameters/projectIdRequiredQueryParam'
      responses:
        '200':
          description: Status retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ServerResponse'
                  - type: object
                    properties:
                      body:
                        $ref: '#/components/schemas/TaskStatus'

    put:
      operationId: putStatusesById
      tags:
        - Statuses
      summary: Update status
      description: Updates status name/category (project manager required).
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskStatusUpdateRequest'
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
        '403':
          description: Insufficient permissions

    delete:
      operationId: deleteStatusesById
      tags:
        - Statuses
      summary: Delete status
      description: |
        Deletes a status and moves tasks to a replacement status.
        Project manager permission required.
      parameters:
        - $ref: '#/components/parameters/idParam'
        - name: project
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Project ID
        - name: replace
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Replacement status ID
      responses:
        '200':
          description: Status deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
        '403':
          $ref: '#/components/responses/CsrfError'
...
  /team-members/project/{id}:
    get:
      operationId: getTeamMembersProjectById
      tags:
        - Team Members
      summary: List members by project
      description: "List members by project. Includes request parameters, authorization expectations, and response details."
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Project members retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
  /team-members/projects/{id}:
    get:
      operationId: getTeamMembersProjectsById
      tags:
        - Team Members
      summary: List projects by team member
      description: "List projects by team member. Includes request parameters, authorization expectations, and response details."
      parameters:
        - $ref: '#/components/parameters/idParam'
        - name: project
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated project IDs filter
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated status IDs filter
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Member projects retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
...
  /reporting/overview/projects/{team_id}:
    get:
      tags:
        - Reporting
      operationId: getReportingOverviewProjectsByTeam
      summary: Overview projects by team
      description: "Overview projects by team. Includes request parameters, authorization expectations, and response details."
      parameters:
        - $ref: '#/components/parameters/teamIdPathParam'
        - $ref: '#/components/parameters/memberQueryParam'
        - $ref: '#/components/parameters/reportingArchived'
      responses:
        '200':
          description: Projects retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
...
  /reporting/overview/project/tasks/{project_id}:
    get:
      tags:
        - Reporting
      operationId: getReportingOverviewProjectTasks
      summary: Project tasks summary
      description: "Project tasks summary. Includes request parameters, authorization expectations, and response details."
      parameters:
        - $ref: '#/components/parameters/projectIdPathParam'
        - name: group
          in: query
          required: false
          schema:
            type: string
            enum: [status, assignee, priority]
      responses:
        '200':
          description: Project task summary retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
...
  /reporting/overview/member/tasks/{team_member_id}:
    get:
      tags:
        - Reporting
      operationId: getReportingOverviewMemberTasks
      summary: Member tasks list
      description: "Member tasks list. Includes request parameters, authorization expectations, and response details."
      parameters:
        - $ref: '#/components/parameters/teamMemberIdPathParam'
        - $ref: '#/components/parameters/projectQueryUuidParam'
        - $ref: '#/components/parameters/onlySingleMemberParam'
        - $ref: '#/components/parameters/durationParam'
        - $ref: '#/components/parameters/dateRangeParam'
        - $ref: '#/components/parameters/reportingArchived'
      responses:
        '200':
          description: Member tasks retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie set after successful login

    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF token obtained from /csrf-token endpoint.
        Required for all POST, PUT, DELETE requests.

    clientPortalToken:
      type: apiKey
      in: header
      name: x-client-token
      description: Client portal authentication token (for client-specific endpoints)
  parameters:
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resource UUID
      example: "550e8400-e29b-41d4-a716-446655440000"
...
    memberQueryParam:
      name: member
      in: query
      required: false
      schema:
        type: string
        format: uuid
      description: Team member ID filter
    teamIdPathParam:
      name: team_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Team ID
    projectIdPathParam:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Project ID
    teamMemberIdPathParam:
      name: team_member_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Team member ID
...
    projectIdRequiredQueryParam:
      name: project_id
      in: query
      required: true
      schema:
        type: string
        format: uuid
      description: Project ID
...
    durationParam:
      name: duration
      in: query
      required: false
      schema:
        type: string
      description: Preset date range key
    dateRangeParam:
      name: date_range
      in: query
      required: false
      schema:
        type: string
      description: Custom date range payload/string
...
    projectQueryUuidParam:
      name: project
      in: query
      required: false
      schema:
        type: string
        format: uuid
      description: Project ID filter
    onlySingleMemberParam:
      name: only_single_member
      in: query
      required: false
      schema:
        type: string
        enum: ["true", "false"]
      description: Restrict to tasks directly assigned to member only
...
    reportingArchived:
      name: archived
      in: query
      required: false
      schema:
        type: string
        enum: ["true", "false"]
      description: Include archived projects/tasks
  schemas:
    ServerResponse:
      type: object
      required:
        - done
        - body
      properties:
        done:
          type: boolean
          description: Indicates if the operation was successful
          example: true
        body:
          description: Response data (type varies by endpoint, null if no data)
          nullable: true
        title:
          type: string
          nullable: true
          description: Optional response title
        message:
          type: string
          nullable: true
          description: Optional message (error message or success notification)
          example: "Operation completed successfully"
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Task UUID
        name:
          type: string
          maxLength: 100
          description: Task name
          example: "Implement user authentication"
        description:
          type: string
          maxLength: 4000
          nullable: true
          description: Task description (supports HTML)
        project_id:
          type: string
          format: uuid
          description: Associated project UUID
        status_id:
          type: string
          format: uuid
          description: Task status UUID
        priority_id:
          type: string
          format: uuid
          nullable: true
          description: Task priority UUID
        assignees:
          type: array
          items:
            type: string
            format: uuid
          description: Array of assigned user UUIDs
        labels:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              color:
                type: string
          description: Task labels with colors
        start_date:
          type: string
          format: date
          nullable: true
          description: Task start date (YYYY-MM-DD)
        end_date:
          type: string
          format: date
          nullable: true
          description: Task due date (YYYY-MM-DD)
        total_minutes:
          type: integer
          description: Estimated time in minutes
          example: 480
        billable:
          type: boolean
          description: Whether the task is billable
          default: false
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
...
    TaskUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 4000
          nullable: true
        status_id:
          type: string
          format: uuid
        priority_id:
          type: string
          format: uuid
          nullable: true
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        billable:
          type: boolean
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
          example: "Website Redesign"
        key:
          type: string
          maxLength: 10
          description: Project key (e.g., "WEB")
          example: "WEB"
        color_code:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          description: Project color in hex format
          example: "#3498db"
        status_id:
          type: string
          format: uuid
          nullable: true
        category_id:
          type: string
          format: uuid
          nullable: true
        folder_id:
          type: string
          format: uuid
          nullable: true
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ProjectUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        key:
          type: string
          maxLength: 10
        color_code:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        status_id:
          type: string
          format: uuid
          nullable: true
        category_id:
          type: string
          format: uuid
          nullable: true
        folder_id:
          type: string
          format: uuid
          nullable: true
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
...
    TaskStatus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color_code:
          type: string
        color_code_dark:
          type: string
          nullable: true
        category_id:
          type: string
          format: uuid
        category_name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
...
    TaskStatusUpdateRequest:
      type: object
      required:
        - name
        - project_id
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        project_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
          nullable: true
...
  responses:
...
    CsrfError:
      description: CSRF token validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ServerResponse'
          example:
            done: false
            body: null
            message: "CSRF token validation failed"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ServerResponse'
          example:
            done: false
            body: null
            message: "Validation error: name is required"
...